services/fastapi-main/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-jose
passlib
python-multipart
httpx
pydantic
pydantic-settings
python-dotenv
pyjwt
python-docx
cairosvg